modulo7: concluido a aula 11Tratamento de erros e useHistory

+++ TreinamentoI-Front-end/modulo7-React/aula-react-gama/src/pages/Repositories/index.js
@@ -1,19 +1,27 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable no-lone-blocks */
 import React, {useState, useEffect} from 'react';
+import {useHistory} from 'react-router-dom';
 
 import * as S from './styled'; // vai importar tudo que estiver dentro de styled para a variável S
 
 function Repositories() {
   const [repositories, setRepositories] = useState([]);
+  const history = useHistory();
 
   useEffect(() => {
     // pegando os dados do localStorage e armazenando em repoName
     let repoName = localStorage.getItem('repoName');
 
-    repoName = JSON.parse(repoName);
-    setRepositories(repoName);
-    // limpando o localStorage depois que ele for setado para buscar os dados que estão armazenados nele
-    localStorage.clear();
+    // se o localStorage não estiver vazio pega os dados, senão volta para a página Home
+    if (repoName !== null) {
+      repoName = JSON.parse(repoName); // realizando o parse do repoName
+      setRepositories(repoName);
+      // limpando o localStorage depois que ele for setado para buscar os dados que estão armazenados nele
+      localStorage.clear();
+    } else {
+      history.push("/"); // retornando para a página principal
+    }
   }, []);
 
   return (
@@ -29,6 +37,7 @@ function Repositories() {
           );
         })}
       </S.List>
+      <S.LinkHome to="/">Voltar</S.LinkHome>
     </S.Container>
   );
 }
